refactor(tests): consolidate video player mocks and remove unused helper files
