--- test/app/features/stories/helpers/fake_video_platform.dart
@@ -1,44 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-import 'package:flutter/widgets.dart';
-import 'package:video_player_platform_interface/video_player_platform_interface.dart';
-
-class FakeVideoPlayerPlatform extends VideoPlayerPlatform {
-  FakeVideoPlayerPlatform() : super();
-
-  @override
-  Future<void> init() async {}
-
-  @override
-  Future<int> create(DataSource dataSource) async => 1;
-
-  @override
-  Future<void> dispose(int textureId) async {}
-
-  @override
-  Future<void> play(int textureId) async {}
-
-  @override
-  Future<void> pause(int textureId) async {}
-
-  @override
-  Future<void> setLooping(int textureId, bool looping) async {}
-
-  @override
-  Future<void> setVolume(int textureId, double volume) async {}
-
-  @override
-  Future<void> setPlaybackSpeed(int textureId, double speed) async {}
-
-  @override
-  Future<void> seekTo(int textureId, Duration position) async {}
-
-  @override
-  Future<Duration> getPosition(int textureId) async => Duration.zero;
-
-  @override
-  Future<void> setMixWithOthers(bool mixWithOthers) async {}
-
-  @override
-  Widget buildView(int textureId) => const SizedBox();
-}
