+++ test/app/features/stories/hooks/use_story_progress_controller_test.dart
@@ -11,8 +11,8 @@ import 'package:ion/app/features/feed/stories/hooks/use_story_video_playback.dar
 import 'package:ion/app/features/feed/stories/providers/story_image_loading_provider.c.dart';
 
 import '../../../../fixtures/stories/story_fixtures.dart';
+import '../../../../mocks.dart';
 import '../helpers/story_test_utils.dart';
-import '../helpers/story_test_video.dart';
 
 class _StoryConsumer extends HookConsumerWidget {
   const _StoryConsumer({required this.post, required this.onCompleted});
